Finished initial http body data class.

+++ src/main/java/com/drkwitht/util/HTTPBody.java
@@ -5,7 +5,7 @@ package com.drkwitht.util;
  */
 public class HTTPBody {
     private HTTPContentType type;
-    private long length;
+    private int length;
     private String content;
 
     public HTTPBody(HTTPContentType typeCode, String contentText) {
